Make FP tolerances configurable for non-fast-math
Before this commit, floating-point arithmetic tolerances written in
the `fastmath.cmake` file were passed to the `fpcmp` command only if
the CMake `TEST_SUITE_FUJITSU_WITH_FAST_MATH` variable is true.
And, those tolerances are tuned for fast-math arithmetic.
For non-fast-math, a fixed value 0x1.0p-51 is used for tolerances.
This value is tuned for IEEE 754 binary64 (double-precision)
arithmetic and is not suitable for IEEE 754 binary32 (single-precision)
arithmetic.

This commit make tolerances configurable by the `fpconfig.cmake` file
for non-fast-math airthmetic. For details, see the comment of the
`llvm_singlesource_fujitsu` function in the top-level `CMakeLists.txt`
file.